create map component and make the main page empty
